Add addError(String messageId, int c) method
This change adds an `addError(String messageId, int c)` method. The
intent of it is for cases like the “illegal character” error, to cite
the offending character in the error message.

So this change also updates the `illegal.character` message property to
add a message argument to it, and updates the `illegal.character` error
call sites to use the `addError(String messageId, int c)` method.

Thus the method expects to be called with a character value for the `c`
int arg, and makes a string from it by first casting that int to a char.